fixed some bugs in compute_group_variances and added basic tests
